ES Linting Question
Hi Wes, if this is the wrong place or way to ask these questions, let me know! I'm noticing that with my ESLint set up, there are quite a few places in this codebase where my ESLinter config is reporting on issues. For example, in this case I had
```
[eslint] Unexpected block statement surrounding arrow body; move the returned value immediately after the `=>`. (arrow-body-style)
```

I'm not sure if it's beyond the scope of this tutorial, but I am interested in your ESLint config setup; I just got mine configured for this project to help highlight these kinds of errors, but mine is seemingly more strict than yours.

There are other places in the code too, but this is more of a general question than me looking to merge this "fix". Here is what I'm using (so far) in my `.eslintrc.json`:
```
{
  "extends": "airbnb-base",
  "env": {
      "es6": true,
      "browser": true
  },
  "rules": {
      "brace-style": [
          "error"
      ],
      "comma-dangle": [
          "error",
          "never"
      ],
      "no-unused-vars": [
          "warn"
      ],
      "no-var": [
          "off"
      ],
      "one-var": [
          "off"
      ],
      "import/newline-after-import": 0,
      "no-underscore-dangle": [
          "error",
          {"allow": ["_id"]}
      ],
      "func-names": ["error", "never"],
      "no-console": 0,
      "consistent-return": 0,
      "arrow-parens": [
          2,
          "as-needed",
          { "requireForBlockBody": false }
        ]
  }
}
```